pyproject.toml
[tool.poetry]
name = "django_sample"
version = "0.1.0"
description = "django sample project"
authors = ["cosmos2 <playnstop.s@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.6"
djangorestframework = "3.12.0"
djangorestframework-camel-case = "^1.2.0"
black = "^21.7b0"
django-extensions = "^3.1.3"
django-taggit = "^1.5.1"
drf-yasg = "^1.20.0"
django-safedelete = "^1.0.0"
isort = "^5.9.3"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.1"

[tool.black]
line-length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
